pyproject.toml
[project]
name = "rag-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-search>=1.1.1",
    "google-search-results>=2.4.2",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-huggingface>=0.3.1",
    "langchain-ollama>=0.3.6",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.6.5",
    "pdf2image>=1.17.0",
    "qdrant-client>=1.15.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.1",
    "surya-ocr>=0.15.4",
    "uvicorn>=0.35.0",
    "websocket-client>=1.8.0",
]
